Use new convenience function to open settings
GS 3.36 introduced a portal-like API for actions like opening settings.
That was done in order to make it possible to ship the new extensions
app in a sandboxed manner, e.g. via flatpak.

While talking to the API directly is a bit cumbersome, luckily there
is a convenience function we can make use of (from GS 3.36.2 on).

Closes #260.

Signed-off-by: Jens Lody <jens@jenslody.de>