Fix 'C.UTF-8' locale.
If the extension is unable to parse the user's locale, it gets
stuck in a retry loop.  There was already a special case for the 'C'
locale to default to 'en'.  This change extends that special case
to cover the 'C.UTF-8' case (which currently breaks the extension) and
all other unknown locales.

Signed-off-by: Craig <cgallek@gmail.com>