Revert to using Adwaita icons since I can't get Breeze icons to display properly
The Breeze icons seem to not be as simple as the Adwaita icons and
display as a white square. I tried using Scour to simplify but it didn't
work and I tried to use Inkscape to convert them to PNGs and that then
worked but the colors are fixed and don't theme correctly as PNGs. This
means the icon set is also once again more limited.