Update es.po
Hi! Spanish native from Argentina here, open to contribute...
I have added some translations, several words used more widely in the Spanish language were also added. And I added critical distinction in Spanish language for the english word 'weather'.

'Weather' in spanish has a distinction and a difference to highlight: as long as we (Spanish natives) talk about weather, we must use the word 'tiempo'.
Tiempo, like the english word 'time', refers to the current weather or the climatic condition in a just moment.
In the other hand, 'Clima' refers to the meteorological properties of a geographical area over a long period of time.

I hope the translation is properly displayed. If not, I am available to contribute with fixes -liking this extension!!.
Best Regards,
Fede.