API Key Array transition and VisualCrossing bones and fix Notify Spam
Transition from specifically named settings for OpenWeatherMap and
WeatherAPI.com for if to use the default and the custom to an array of
keys where an empty string indicates to use the default key and the
index in the array is equivalent to the provider's position in the enum
(minus one because of the default/"Adaptive" mode).

Bones for fetching VisualCrossing data are there however data is not yet
processed.

Notification spam when provider is full no longer happens as the
1-second retry no longer presists as the refresh rate.