fix repeat of day in forecast view
The previous code would repeat the forecast for a timeslot when the forecast data would roll over from one day into the next. This code fixes that issue by letting the loop go onto the next iteration to pick up the data from the first index of the next day.